+++ test/ClusteredLighting.cpp
@@ -44,8 +44,8 @@ static void openGfxWindow(eve::window::Window *&win, Graphics *&gfx, int w = 320
     REQUIRE(win != nullptr);
     REQUIRE(gfx != nullptr);
     eve::window::WindowSettings s;
-    s.width = w;
-    s.height = h;
+    s.width  = static_cast<uint16_t>(w);
+    s.height = static_cast<uint16_t>(h);
     s.centered = true;
     REQUIRE(win->setWindowSettings(s));
 }
